se agrego colaGenerica, pilaPremios/castigos se elimino metodo repetido
